fix: Wrap custom approximate with Debug/NUnit checks
Due to the Mathf.Approximately checking for the smallest difference, tests with values of 0.2 and0.2000003 are failing. Accumulations of floating point imprecisions with calculations are to be expected and thus require a larger range of impression. The changes help avoid test failure while identifying such instances and removing the extra check when not unit testing.